Incoporate code review feedback
